requirements.txt
colorama==0.4.6
iniconfig==2.3.0
packaging==26.2
pluggy==1.6.0
Pygments==2.20.0
pytest==9.0.3
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.4.22
charset-normalizer==3.4.7
colorama==0.4.6
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
iniconfig==2.3.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.2.17
langchain-core==1.3.3
langchain-protocol==0.0.15
langgraph==1.1.10
langgraph-checkpoint==4.0.3
langgraph-prebuilt==1.0.13
langgraph-sdk==0.3.14
langsmith==0.8.2
orjson==3.11.9
ormsgpack==1.12.2
packaging==26.2
pluggy==1.6.0
pydantic==2.13.4
pydantic_core==2.46.4
Pygments==2.20.0
pytest==9.0.3
PyYAML==6.0.3
requests==2.33.1
requests-toolbelt==1.0.0
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.7.0
uuid_utils==0.14.1
xxhash==3.7.0
zstandard==0.25.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
appdirs==1.4.4
attrs==26.1.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
dataclasses-json==0.6.7
datasets==4.8.5
dill==0.4.1
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.18.0
filelock==3.29.0
frozenlist==1.8.0
fsspec==2026.2.0
greenlet==3.5.0
h11==0.16.0
hf-xet==1.5.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.14.0
idna==3.13
iniconfig==2.3.0
instructor==1.15.1
Jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.2.17
langchain-classic==1.0.7
langchain-community==0.4.1
langchain-core==1.3.3
langchain-openai==1.2.1
langchain-protocol==0.0.15
langchain-text-splitters==1.1.2
langgraph==1.1.10
langgraph-checkpoint==4.0.3
langgraph-prebuilt==1.0.13
langgraph-sdk==0.3.14
langsmith==0.8.2
markdown-it-py==4.2.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
multidict==6.7.1
multiprocess==0.70.19
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.4
openai==2.36.0
orjson==3.11.9
ormsgpack==1.12.2
packaging==26.2
pandas==3.0.2
pillow==12.2.0
pluggy==1.6.0
propcache==0.4.1
pyarrow==24.0.0
pydantic==2.13.4
pydantic-settings==2.14.0
pydantic_core==2.46.4
Pygments==2.20.0
pytest==9.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
ragas==0.4.3
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
rich==14.3.4
scikit-network==0.33.5
scipy==1.17.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typer==0.25.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.7.0
uuid_utils==0.14.1
xxhash==3.7.0
yarl==1.23.0
zstandard==0.25.0
